+++ crates/store/src/db/mod.rs
@@ -11,7 +11,14 @@ use miden_objects::account::AccountId;
 use miden_objects::asset::Asset;
 use miden_objects::block::{BlockHeader, BlockNoteIndex, BlockNumber, ProvenBlock};
 use miden_objects::crypto::merkle::SparseMerklePath;
-use miden_objects::note::{NoteDetails, NoteId, NoteInclusionProof, NoteMetadata, Nullifier};
+use miden_objects::note::{
+    NoteDetails,
+    NoteId,
+    NoteInclusionProof,
+    NoteMetadata,
+    NoteScript,
+    Nullifier,
+};
 use miden_objects::transaction::TransactionId;
 use tokio::sync::oneshot;
 use tracing::{info, info_span, instrument};
@@ -535,6 +542,14 @@ impl Db {
     ) -> Result<(BlockNumber, Vec<AccountVaultValue>)> {
         self.transact("account vault sync", move |conn| {
             queries::select_account_vault_assets(conn, account_id, block_from, block_to)
+        })
+        .await
+    }
+
+    /// Returns the script for a note by its root.
+    pub async fn select_note_script_by_root(&self, root: Word) -> Result<Option<NoteScript>> {
+        self.transact("note script by root", move |conn| {
+            queries::select_note_script_by_root(conn, root)
         })
         .await
     }